node-pigz: Don't hardcode path to pigz binary
Instead of executing the binary at the hard-coded path '/usr/local/bin/pigz', we let the system find the 'pigz' binary (using $PATH, etc.)

We additionally check that the system can find and execute the pigz executable at module init. If it fails, we redirect deflate() to a method which will simply throw an Error about not finding pigz, and returns.